Rename sample doc files
